requirements.txt
# read a text from image
easyocr

# autoclicker
pyautogui

# work with images (of course)
pillow

# screenshots
mss
